Pulled out FileTime class to j2objc/nio/file/attribute. Changed relevant licenses. Modified implementation for FileTime class.
