+++ jre_emul/android/platform/libcore/ojluni/src/main/java/java/util/zip/ZipEntry.java
@@ -27,14 +27,14 @@
 package java.util.zip;
 
 /* J2ObjC Added: import FileTime class to avoid java.nio.file.attribute.FileTime */
-import static com.google.j2objc.util.ZipUtils.*;
+import com.google.j2objc.nio.file.attribute.FileTime;
 import java.nio.charset.StandardCharsets;
 /* J2ObjC Removed: avoid using Java FileTime.
 import java.nio.file.attribute.FileTime; */
 import java.util.Objects;
 import java.util.concurrent.TimeUnit;
-
 import static java.util.zip.ZipConstants64.*;
+import static java.util.zip.ZipUtils.*;
 
 /**
  * This class is used to represent a ZIP file entry.
